(feat): add PATCH request to rename shortcut
The resource drives/${User}/root/children adds the name of the library and site to the name of the shortcut.

That is not likely to be what people want when they pass a shortcut name to the script.

This commit adds a PATCH request to users/${User}/drive/items/${ItemId} to rename the shortcut after it was created, so it matches what was passed to the script.