(update): change drive endpoint
There is an issue reported on https://github.com/innovara/odsc/issues/1

Although the origin is not clear, a user reports that using endpoint /users/{idOrUserPrincipalName}/drive instead of /drives/{idOrUserPrincipalName} fixed the issue for them.

While it is possible that it wasn't the case and these drives throwing server error 500 were in fact new drives that were not provisioned yet, using /users/{idOrUserPrincipalName}/drive on those throws a '404 Not Found' error which might be more telling and helpful than '500 Internal Server Error'.

Change /drives/{idOrUserPrincipalName} to /users/{idOrUserPrincipalName}/drive

Signed-off-by: Manuel Fombuena <mfombuena@innovara.tech>