(update): ensure resource is remoteItem on Delete
Especially when used programmatically, Remove-odsc can encounter an existing resource that is not a shortcut but it is named as the resource we want to delete.

The reasonable expectation in this regard is that the non-shortcut resource is not deleted.

Add logic to check for remoteItem resource type before Delete.

Signed-off-by: Manuel Fombuena <mfombuena@innovara.tech>